+++ packages/video_player/video_player_platform_interface/test/video_player_platform_interface_test.dart
@@ -19,4 +19,28 @@ void main() {
       throwsUnimplementedError,
     );
   });
+
+  test('default implementation getAudioTracks throws unimplemented', () async {
+    await expectLater(
+      () => initialInstance.getAudioTracks(1),
+      throwsUnimplementedError,
+    );
+  });
+
+  test('default implementation selectAudioTrack throws unimplemented',
+      () async {
+    await expectLater(
+      () => initialInstance.selectAudioTrack(1, 'trackId'),
+      throwsUnimplementedError,
+    );
+  });
+
+  test(
+      'default implementation isAudioTrackSupportAvailable throws unimplemented',
+      () async {
+    await expectLater(
+      () => initialInstance.isAudioTrackSupportAvailable(),
+      throwsUnimplementedError,
+    );
+  });
 }