+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 6.5.0
 
-* Adds `VideoAudioTrack` class and `getAudioTracks()`, `selectAudioTrack()` methods to platform interface for audio track management.
+* Adds `VideoAudioTrack` class and `getAudioTracks()`, `selectAudioTrack()`, `isAudioTrackSupportAvailable()` methods to platform interface for audio track management.
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 
 ## 6.4.0